separate panderas, load_dataset, log functions from general.py
